fix(p1+errorx): fix 8 P1 issues and introduce errorx error handling
P1 fixes:
- Remove dead CLI flags -duplicate-singular and -allow-json-unknown-fields
- Fix fieldNumber<<3 integer overflow on 32-bit platforms (wire.go, marshal.go)
- Reject bytes items required constraint at parser layer
- Use ShouldBindJSON instead of ShouldBind in generated HTTP handlers
- Fix MatchPattern nil dereference on type assertion failure
- Replace UNKNOWN sentinel with panic in GoScalarType and resolveGoType
- Add gcode-generated method names to reservedFieldNames
- Add collision detection for enumIndex (cross-file same-name enums)

errorx integration:
- Introduce github.com/pinealctx/x/errorx as base error layer
- Define ParseError/ConfigError/AppError domain sentinel types
- Migrate constraint violation errors to errorx.NewSentinelf[parserTag]
- Migrate config validation errors to fixed ConfigError sentinels
- Migrate app coordination errors to errorx.NewSentinelf[appTag]
- Add BizCode type and CodeOK/CodeDefaultErr/CodeValidationErr constants to httpruntime
- Support errorx.Error[BizCode] in ErrResponse with chain traversal
- Upgrade all error-path tests from strings.Contains to errors.Is/As
- Add config_test.go with errors.Is coverage for all config sentinels
- Replace hardcoded 0/500/400 with httpruntime constants in rpc_test.go

+++ runtime/wire.go
@@ -18,7 +18,7 @@ const (
 
 // AppendTag appends a protobuf field tag (field_number << 3 | wire_type).
 func AppendTag(b []byte, fieldNumber int, wireType int) []byte {
-	return AppendVarint(b, uint64(fieldNumber<<3|wireType))
+	return AppendVarint(b, uint64(fieldNumber)<<3|uint64(wireType))
 }
 
 // AppendVarint appends a base-128 varint-encoded uint64.
@@ -98,7 +98,7 @@ func SizeVarint(v uint64) int {
 
 // SizeTag returns the number of bytes needed to encode a field tag.
 func SizeTag(fieldNumber int) int {
-	return SizeVarint(uint64(fieldNumber << 3))
+	return SizeVarint(uint64(fieldNumber) << 3)
 }
 
 // DecodeZigZag32 decodes a ZigZag-encoded uint64 back to int32.
